Refactored account-db.js into separate files in accounts/
