+++ src/scripts/reset-password.js
@@ -1,4 +1,5 @@
-import { needsBootstrap, bootstrap, changePassword } from '../account-db.js';
+import { needsBootstrap, bootstrap } from '../accounts/index.js';
+import { changePassword } from '../accounts/password.js';
 import { promptPassword } from '../util/prompt.js';
 
 if (needsBootstrap()) {
