+++ src/services/secrets-service.js
@@ -2,7 +2,7 @@ import createDebug from 'debug';
 import fs from 'node:fs';
 import config, { sqlDir } from '../load-config.js';
 import { join } from 'node:path';
-import getAccountDb from '../account-db.js';
+import getAccountDb from '../accounts/index.js';
 
 /**
  * An enum of valid secret names.
